Only track horizontal vehicle motion
At the start of simulation, the vehicle is actually hovering slightly
above the platform; this prevents the simulator from erroneously
inferring initial conditions under which the car intersects the
platform.  This does, however, mean that the car has to fall a couple
inches onto the platform when the simulation begins, so if you record
the absolute speed of the car, you’ll get a weird spike at the start of
the simulation.  This commit fixes that issue by confining speed
measurements to the xy plane.